fixed bug in file cleanup

+++ winforms/src/toga_winforms/widgets/webview.py
@@ -185,7 +185,9 @@ class WebView(Widget):
             )
 
     def winforms_navigation_completed(self, sender, args):
-        if os.path.exists(self._large_content_filepath):
+        if self._large_content_filepath and os.path.exists(
+            self._large_content_filepath
+        ):
             os.remove(self._large_content_filepath)
             self._large_content_filepath = None
         self.interface.on_webview_load()
